LPD-52040 Extract ManagementBar to reuse it and create StructureBuilderManagementBar
