+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/structure_builder/components/ManagementBar.tsx
@@ -5,28 +5,15 @@
 
 import ClayIcon from '@clayui/icon';
 import ClayLink from '@clayui/link';
-import {openConfirmModal} from '@liferay/layout-js-components-web';
-import {ManagementToolbar, openToast} from 'frontend-js-components-web';
+import {ManagementToolbar} from 'frontend-js-components-web';
 import React from 'react';
 
-import {useSelector, useStateDispatch} from '../contexts/StateContext';
-import selectHistory from '../selectors/selectHistory';
-import selectStructureERC from '../selectors/selectStructureERC';
-import selectStructureFields from '../selectors/selectStructureFields';
-import selectStructureId from '../selectors/selectStructureId';
-import selectStructureLabel from '../selectors/selectStructureLabel';
-import selectStructureLocalizedLabel from '../selectors/selectStructureLocalizedLabel';
-import selectStructureName from '../selectors/selectStructureName';
-import selectStructureSpaces from '../selectors/selectStructureSpaces';
-import selectStructureStatus from '../selectors/selectStructureStatus';
-import StructureService from '../services/StructureService';
-import {useValidate} from '../utils/validation';
-import AsyncButton from './AsyncButton';
+type Props = {
+	children: React.ReactNode;
+	title: string;
+};
 
-export default function ManagementBar() {
-	const label = useSelector(selectStructureLocalizedLabel);
-	const status = useSelector(selectStructureStatus);
-
+export default function ManagementBar({children, title}: Props) {
 	return (
 		<ManagementToolbar.Container className="border">
 			<ManagementToolbar.ItemList className="c-gap-3" expand>
@@ -42,9 +29,7 @@ export default function ManagementBar() {
 
 				<ManagementToolbar.Item className="nav-item-expand">
 					<h2 className="font-weight-semi-bold m-0 text-5">
-						{status === 'published'
-							? label
-							: Liferay.Language.get('new-structure')}
+						{title}
 					</h2>
 				</ManagementToolbar.Item>
 
@@ -57,201 +42,8 @@ export default function ManagementBar() {
 					</ClayLink>
 				</ManagementToolbar.Item>
 
-				{status !== 'published' ? (
-					<ManagementToolbar.Item>
-						<SaveButton />
-					</ManagementToolbar.Item>
-				) : null}
-
-				<ManagementToolbar.Item>
-					<PublishButton />
-				</ManagementToolbar.Item>
+				{children}
 			</ManagementToolbar.ItemList>
 		</ManagementToolbar.Container>
-	);
-}
-
-function SaveButton() {
-	const dispatch = useStateDispatch();
-	const validate = useValidate();
-
-	const erc = useSelector(selectStructureERC);
-	const fields = useSelector(selectStructureFields);
-	const label = useSelector(selectStructureLabel);
-	const localizedLabel = useSelector(selectStructureLocalizedLabel);
-	const name = useSelector(selectStructureName);
-	const spaces = useSelector(selectStructureSpaces);
-	const status = useSelector(selectStructureStatus);
-	const structureId = useSelector(selectStructureId);
-
-	const onSave = async () => {
-		const valid = validate();
-
-		if (!valid) {
-			return;
-		}
-
-		try {
-			if (status === 'new') {
-				const {id} = await StructureService.createStructure({
-					erc,
-					fields,
-					label,
-					name,
-					spaces,
-				});
-
-				dispatch({id, type: 'create-structure'});
-			}
-			else {
-				await StructureService.updateStructure({
-					erc,
-					fields,
-					id: structureId,
-					label,
-					name,
-					spaces,
-				});
-
-				dispatch({type: 'clear-error'});
-			}
-
-			openToast({
-				message: Liferay.Util.sub(
-					Liferay.Language.get('x-was-saved-successfully'),
-					localizedLabel
-				),
-				type: 'success',
-			});
-		}
-		catch (error) {
-			const {message} = error as Error;
-
-			dispatch({
-				error:
-					message ||
-					Liferay.Language.get(
-						'an-unexpected-error-occurred-while-saving-or-publishing-the-structure'
-					),
-				type: 'set-error',
-			});
-		}
-	};
-
-	return (
-		<AsyncButton
-			displayType="secondary"
-			label={Liferay.Language.get('save')}
-			onClick={onSave}
-		/>
-	);
-}
-
-function PublishButton() {
-	const dispatch = useStateDispatch();
-	const validate = useValidate();
-
-	const erc = useSelector(selectStructureERC);
-	const fields = useSelector(selectStructureFields);
-	const history = useSelector(selectHistory);
-	const label = useSelector(selectStructureLabel);
-	const localizedLabel = useSelector(selectStructureLocalizedLabel);
-	const name = useSelector(selectStructureName);
-	const spaces = useSelector(selectStructureSpaces);
-	const status = useSelector(selectStructureStatus);
-	const structureId = useSelector(selectStructureId);
-
-	const onPublish = async () => {
-		const valid = validate();
-
-		if (!valid) {
-			return;
-		}
-
-		if (history.deletedFields) {
-			if (
-				!(await openConfirmModal({
-					buttonLabel: Liferay.Language.get('publish'),
-					center: true,
-					status: 'danger',
-					text: Liferay.Language.get(
-						'you-removed-one-or-more-fields-from-the-structure'
-					),
-					title: Liferay.Language.get('publish-structure-changes'),
-				}))
-			) {
-				return;
-			}
-		}
-
-		try {
-			if (status === 'new') {
-				const {id} = await StructureService.createStructure({
-					erc,
-					fields,
-					label,
-					name,
-					spaces,
-				});
-
-				await StructureService.publishStructure({id});
-
-				dispatch({id, type: 'publish-structure'});
-			}
-			else if (status === 'draft') {
-				await StructureService.updateStructure({
-					erc,
-					fields,
-					id: structureId,
-					label,
-					name,
-					spaces,
-				});
-
-				await StructureService.publishStructure({id: structureId});
-
-				dispatch({type: 'publish-structure'});
-			}
-			else if (status === 'published') {
-				await StructureService.updateStructure({
-					erc,
-					fields,
-					id: structureId,
-					label,
-					name,
-					spaces,
-				});
-
-				dispatch({type: 'publish-structure'});
-			}
-
-			openToast({
-				message: Liferay.Util.sub(
-					Liferay.Language.get('x-was-published-successfully'),
-					localizedLabel
-				),
-				type: 'success',
-			});
-		}
-		catch (error) {
-			const {message} = error as Error;
-
-			dispatch({
-				error:
-					message ||
-					Liferay.Language.get(
-						'an-unexpected-error-occurred-while-saving-or-publishing-the-structure'
-					),
-				type: 'set-error',
-			});
-		}
-	};
-
-	return (
-		<AsyncButton
-			displayType="primary"
-			label={Liferay.Language.get('publish')}
-			onClick={onPublish}
-		/>
 	);
 }
